fix(deno-lint): Support files.include from .denolint.json
Complete the existing support of files.exclude.

Once files.include is provided, the current directory will not be
added to the source directories by default.

BREAKING CHANGE: If you used files.include in your .denolint.json
to list a subset of the subdirectories, but you expected denolint
to search all subdirectories in the current directory, this will
not happen any more. If you use a config file with files.include,
only directories listed by files.include will be searched.