fix(deno-lint): Support directories to scan on the command line
Allow limiting the scan to selected directories. If no directory
is provided, the current directory will be scanned, as it was.

BREAKING CHANGE: If you passed a single directory to denolint to
specify the project root, it will not work any more. You can pass
the root directory using the argument -r,--root. However, it was
used only if neither .denolintignore nor .eslintignore were found
and was passed instead of either of them to the walker. Because
the root directory was no file path, I think that it was useless.
Nevertheless, I implemented the previous functionality using the
argument -r,--root, which might be worth for other purposes later.